requirements.txt
asgiref==3.2.5
certifi==2019.11.28
chardet==3.0.4
coverage==5.0.4
Django==3.0.4
django-admin-interface==0.12.1
django-ckeditor==5.9.0
django-cleanup==4.0.0
django-colorfield==0.2.1
django-crispy-forms==1.9.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-image-optimizer==1.0.0
django-js-asset==1.2.2
django-taggit==1.2.0
idna==2.9
Pillow==7.0.0
pipfile==0.0.2
pytz==2019.3
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
tinify==1.5.1
toml==0.10.0
urllib3==1.25.8
